refactor structure search functions and endpoints to remove duplicated code blocks
