+++ src/command_line/search/mod.rs
@@ -1,4 +1,5 @@
 pub mod basic_search;
+pub mod cli_structure_search;
 pub mod identity_search;
 pub mod substructure_search;
 pub mod superstructure_search;
