+++ src/rest_api/api/search/mod.rs
@@ -7,8 +7,5 @@ pub use search_management::*;
 mod identity_search;
 pub use identity_search::*;
 
-mod substructure_search;
-pub use substructure_search::*;
-
-mod superstructure_search;
-pub use superstructure_search::*;
+mod structure_search;
+pub use structure_search::*;
